refactor(Button): useButtonWrapperをButtonWrapperに統合
useButtonWrapperはButtonWrapperからしか利用されていなかったため、
CLAUDE.mdの規約（単一ファイルからしか使われないcustom hookは切り出さない）に従い
ButtonWrapper.tsxに直接実装した。振る舞い・DOM構造・出力classNameに変更はない。
